Moved topLevelCategory and FeaturedCategory into shared model
