+++ app/src/main/java/com/sthoray/allright/data/model/main/FeatureCategoriesResponse.kt
@@ -1,15 +1,17 @@
 package com.sthoray.allright.data.model.main
 
+import com.sthoray.allright.data.model.shared.Category
+
 /**
  * Model representing a category feature panel.
  *
  * AllGoods are using integer object names for each category. This can be
- * represented as a map containing indices and [FeatureCategory]s. It may
+ * represented as a map containing indices and [Category]s. It may
  * be useful to convert [categories] into a List: E.g.
  * myObj.categories.values.toList()
  *
- * @property categories a map with [FeatureCategory] as values
+ * @property categories a map with [Category] as values
  */
 data class FeatureCategoriesResponse(
-    val categories: Map<Int, FeatureCategory>
+    val categories: Map<Int, Category>
 )
