+++ app/src/main/java/com/sthoray/allright/data/model/shared/Category.kt
@@ -1,11 +1,11 @@
-package com.sthoray.allright.data.model.main
+package com.sthoray.allright.data.model.shared
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
 
 /**
- * Model representing a featured category.
+ * Model representing a category.
  *
  * @property id the category ID
  * @property name the name of the category
@@ -14,14 +14,14 @@ import com.google.gson.annotations.SerializedName
  * @property imagePath the url path to an associated image
  * @property webPath the url path to browse the category in a web browser
  */
-data class FeatureCategory(
+data class Category(
     val id: Int,
     val name: String,
     val icon: String,
     @SerializedName("listing_count")
     val listingCount: Int,
     @SerializedName("image")
-    val imagePath: String,
+    val imagePath: String?,
     @SerializedName("url")
-    val webPath: String
+    val webPath: String?
 )
