+++ app/src/main/java/com/sthoray/allright/ui/main/viewmodel/MainViewModel.kt
@@ -3,7 +3,7 @@ package com.sthoray.allright.ui.main.viewmodel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.sthoray.allright.data.model.browse.TopLevelCategory
+import com.sthoray.allright.data.model.shared.Category
 import com.sthoray.allright.data.model.main.FeatureCategoriesResponse
 import com.sthoray.allright.data.repository.AppRepository
 import com.sthoray.allright.utils.Resource
@@ -25,7 +25,7 @@ class MainViewModel(
     val featureCategories: MutableLiveData<Resource<FeatureCategoriesResponse>> = MutableLiveData()
 
     /** Top level categories response. */
-    val topLevelCategories: MutableLiveData<Resource<List<TopLevelCategory>>> = MutableLiveData()
+    val topLevelCategories: MutableLiveData<Resource<List<Category>>> = MutableLiveData()
 
     /** Make network requests on initialisation. */
     init {
@@ -57,8 +57,8 @@ class MainViewModel(
     }
 
     private fun handleTopLevelCategoriesResponse(
-        response: Response<List<TopLevelCategory>>
-    ): Resource<List<TopLevelCategory>> {
+        response: Response<List<Category>>
+    ): Resource<List<Category>> {
         if (response.isSuccessful) {
             response.body()?.let { responseBody ->
                 return Resource.Success(responseBody)
